pyproject.toml
[tool.poetry]
name = "MrfmSim"
version = "0.2.0"
description = "Simulate magnetic resonance force microscopy experiments."
authors = ["Peter Sun <hs859@cornell.edu>", "John Marohn <jam99@cornell.edu>"]
maintainers = ["Peter Sun <hs859@cornell.edu>"]
repository = "https://www.github.com/Marohn-Group/mrfmsim"
# adds all subpackages
packages = [{ include = "mrfmsim" }]
readme = "README.rst"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.10"
numpy = ">=1.18"
pyyaml = "^6.0"
mmodel = { git = "https://www.github.com/Marohn-Group/mmodel", branch = "develop" }
click = ">=8.1.3"
numba = ">=0.58.0"
pint = { version = ">=0.22", optional = true }
tox = { version = ">=3.25.0", optional = true }
tox-conda = { version = ">=0.9.2", optional = true } 
pytest = { version = ">=7.1.1", optional = true }
pytest-cov = { version = ">=4.0.0", optional = true }
sphinx = { version = "^6.2.1", optional = true }
sphinx-book-theme = { version = "^1.0.1", optional = true }
nbsphinx = { version = "^0.9.3", optional = true }

[tool.poetry.scripts]
mrfmsim = "mrfmsim.cli:cli"

[tool.poetry.extras]
test = ["tox", "pytest", "pytest-cov", "pint"]
docs = ["nbsphinx", "sphinx", "sphinx_book_theme"]

[tool.pytest.ini_options]
filterwarnings = [
    # note the use of single quote below to denote "raw" strings in TOML
    # the escape sequence is necessary in graphviz dot string
    # used to left align node text
    'ignore:invalid escape sequence \\l:DeprecationWarning',
    # The parallel method of numba might not be used in test
    # ignore the warning
    'ignore::numba.NumbaPerformanceWarning',
    # the plot plugin's mayavi package has a deprecation warning
    'ignore::DeprecationWarning',
]

[tool.coverage.run]
# pytest coverage
source_pkgs = ['mrfmsim']
omit = ['tests/*']

[tool.tox]
# for testing different python environments

legacy_tox_ini = """
[tox]
requires = tox-conda
envlist = py310, py311, coverage, docs
isolated_build = True

[testenv:py310]
deps = 
    pytest >= 7.1.1
    pint >= 0.22
commands = pytest

[testenv:py311]
deps = 
    pytest >= 7.1.1
    pint >= 0.22
commands = pytest

[testenv:py312]
deps = 
    pytest >= 7.1.1
    pint >= 0.22
commands = pytest

[testenv:coverage]
deps = 
    pytest >= 7.1.1
    pytest-cov >= 3.0.0
    pint >= 0.22
commands =
    coverage run --source=tests -m pytest
    coverage html
    coverage report

[testenv:docs]
deps = 
    sphinx == 6.2.1
    sphinx-book-theme == 1.0.1
    nbsphinx == 0.9.3
commands =
    # sphinx-build -W -b html ./docs ./docs/_build/html
    sphinx-build -W -b html -d "{toxworkdir}/docs" docs "{toxinidir}/docs/_build/html"
"""
